+++ mlxtend/_base/_regressor.py
@@ -7,6 +7,7 @@
 # License: BSD 3 clause
 
 import numpy as np
+from time import time
 
 
 class _Regressor(object):
@@ -44,6 +45,7 @@ class _Regressor(object):
         self._check_target_array(y)
         if hasattr(self, 'self.random_seed') and self.random_seed:
             self._rgen = np.random.RandomState(self.random_seed)
+        self._init_time = time()
         self._fit(X=X, y=y, init_params=init_params)
         self._is_fitted = True
         return self